pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "gerd"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "ray",
    "torch",
    "tqdm",
    "pyyaml",
]

[project.scripts]
gerd = "gerd.__main__:cli"
